requirements.txt
# Core Dependencies
# Note: Python 3.10+ is required (check with: python --version)

# Browser Automation
playwright>=1.40.0

# AI/ML Libraries
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU available

# Document Parsing
PyMuPDF>=1.23.0  # PDF parsing
python-docx>=1.1.0  # DOCX parsing

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.1

# Database
sqlalchemy>=2.0.0

# Security/Encryption
cryptography>=41.0.0

# UI (System Tray)
PyQt5>=5.15.10

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Logging
colorlog>=6.8.0

